// src/app/shared/models/employee-skill.ts
export class EmployeeSkill {
  // Category: {Id: 0, Name: "Technology", Code: "TECH", CssClass: "skill-tech"}
  Experience: number;
  ExperienceNormalized: number;
  ExperiencePoints: number;
  Id: number;
  Investment: number;
  IsCapability: boolean;
  IsDisabled: boolean;
  Name: string;
  // Sources: [{Source: 1, SourceText: "Employee Profile", Level: 0, Experience: 80, Description: "9x WCF"},…]
}